// src/lib/saleor/generated/graphql.ts
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID'];
};
// ...
export type ProductVariant = Node &
  ObjectWithMetadata & {
    /** List of attributes assigned to this variant. */
    attributes: Array<SelectedAttribute>;
    /** Channel given to retrieve this product variant. Also used by federation gateway to resolve this object in a federated query. */
    channel?: Maybe<Scalars['String']>;
    /**
     * List of price information in channels for the product.
     *
     * Requires one of the following permissions: AUTHENTICATED_APP, AUTHENTICATED_STAFF_USER.
     */
    channelListings?: Maybe<Array<ProductVariantChannelListing>>;
    created: Scalars['DateTime'];
    /**
     * Digital content for the product variant.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    digitalContent?: Maybe<DigitalContent>;
    /**
     * External ID of this product.
     *
     * Added in Saleor 3.10.
     */
    externalReference?: Maybe<Scalars['String']>;
    id: Scalars['ID'];
    /**
     * List of images for the product variant.
     * @deprecated This field will be removed in Saleor 4.0. Use the `media` field instead.
     */
    images?: Maybe<Array<ProductImage>>;
    /** Gross margin percentage value. */
    margin?: Maybe<Scalars['Int']>;
    /** List of media for the product variant. */
    media?: Maybe<Array<ProductMedia>>;
    /** List of public metadata items. Can be accessed without permissions. */
    metadata: Array<MetadataItem>;
    /**
     * A single key from public metadata.
     *
     * Tip: Use GraphQL aliases to fetch multiple keys.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    metafield?: Maybe<Scalars['String']>;
    /**
     * Public metadata. Use `keys` to control which fields you want to include. The default is to include everything.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    metafields?: Maybe<Scalars['Metadata']>;
    name: Scalars['String'];
    /**
     * Preorder data for product variant.
     *
     * Added in Saleor 3.1.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    preorder?: Maybe<PreorderData>;
    /** Lists the storefront variant's pricing, the current price and discounts, only meant for displaying. */
    pricing?: Maybe<VariantPricingInfo>;
    /** List of private metadata items. Requires staff permissions to access. */
    privateMetadata: Array<MetadataItem>;
    /**
     * A single key from private metadata. Requires staff permissions to access.
     *
     * Tip: Use GraphQL aliases to fetch multiple keys.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    privateMetafield?: Maybe<Scalars['String']>;
    /**
     * Private metadata. Requires staff permissions to access. Use `keys` to control which fields you want to include. The default is to include everything.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    privateMetafields?: Maybe<Scalars['Metadata']>;
    product: Product;
    /** Quantity of a product available for sale in one checkout. Field value will be `null` when no `limitQuantityPerCheckout` in global settings has been set, and `productVariant` stocks are not tracked. */
    quantityAvailable?: Maybe<Scalars['Int']>;
    quantityLimitPerCustomer?: Maybe<Scalars['Int']>;
    /**
     * Total quantity ordered.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    quantityOrdered?: Maybe<Scalars['Int']>;
    /**
     * Total revenue generated by a variant in given period of time. Note: this field should be queried using `reportProductSales` query as it uses optimizations suitable for such calculations.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    revenue?: Maybe<TaxedMoney>;
    sku?: Maybe<Scalars['String']>;
    /**
     * Stocks for the product variant.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS, MANAGE_ORDERS.
     */
    stocks?: Maybe<Array<Stock>>;
    trackInventory: Scalars['Boolean'];
    /** Returns translated product variant fields for the given language code. */
    translation?: Maybe<ProductVariantTranslation>;
    updatedAt: Scalars['DateTime'];
    weight?: Maybe<Weight>;
  };
// ...
export type GetCategoryProductsBySlugQueryVariables = Exact<{
  slug: Scalars['String'];
}>;

export type GetCategoryProductsBySlugQuery = {
  category?: {
    products?: {
      edges: Array<{
        node: {
          id: string;
          slug: string;
          name: string;
          isAvailableForPurchase?: boolean | null;
          description?: string | null;
          seoTitle?: string | null;
          seoDescription?: string | null;
          updatedAt: string;
          pricing?: {
            priceRange?: {
              start?: { gross: { currency: string; amount: number } } | null;
              stop?: { gross: { currency: string; amount: number } } | null;
            } | null;
          } | null;
          media?: Array<{ url: string; type: ProductMediaType; alt: string }> | null;
          collections?: Array<{ name: string }> | null;
          variants?: Array<{
            id: string;
            name: string;
            attributes: Array<{
              attribute: {
                slug?: string | null;
                name?: string | null;
                choices?: { edges: Array<{ node: { name?: string | null } }> } | null;
              };
              values: Array<{ name?: string | null }>;
            }>;
            pricing?: { price?: { gross: { currency: string; amount: number } } | null } | null;
          }> | null;
        };
      }>;
    } | null;
  } | null;
};
// ...
export type GetCollectionProductsBySlugQueryVariables = Exact<{
  slug: Scalars['String'];
}>;
// ...
export type MenuItemFragment = {
  id: string;
  name: string;
  url?: string | null;
  collection?: { slug: string; products?: { totalCount?: number | null } | null } | null;
  category?: { slug: string; products?: { totalCount?: number | null } | null } | null;
  page?: { slug: string } | null;
};

export type GetMenuBySlugQueryVariables = Exact<{
  slug: Scalars['String'];
}>;
// ...
export type GetPageBySlugQueryVariables = Exact<{
  slug: Scalars['String'];
}>;
// ...
export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: { hash: string }) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
// ...
export const GetCategoryProductsBySlugDocument = new TypedDocumentString(`
    query GetCategoryProductsBySlug($slug: String!) {
  category(slug: $slug) {
    products(channel: "default-channel", first: 100) {
      edges {
        node {
          id
          slug
          name
          isAvailableForPurchase
          description
          seoTitle
          seoDescription
          pricing {
            priceRange {
              start {
                gross {
                  currency
                  amount
                }
              }
              stop {
                gross {
                  currency
                  amount
                }
              }
            }
          }
          media {
            url(size: 2160)
            type
            alt
          }
          collections {
            name
          }
          updatedAt
          variants {
            ...Variant
          }
        }
      }
    }
  }
}
    fragment Variant on ProductVariant {
  id
  name
  attributes {
    attribute {
      slug
      name
      choices(first: 100) {
        edges {
          node {
            name
          }
        }
      }
    }
    values {
      name
    }
  }
  pricing {
    price {
      gross {
        currency
        amount
      }
    }
  }
}`) as unknown as TypedDocumentString<
  GetCategoryProductsBySlugQuery,
  GetCategoryProductsBySlugQueryVariables
>;
// ...
export const GetCollectionProductsBySlugDocument = new TypedDocumentString(`
    query GetCollectionProductsBySlug($slug: String!) {
  collection(channel: "default-channel", slug: $slug) {
    products(first: 100) {
      edges {
        node {
          id
          slug
          name
          isAvailableForPurchase
          description
          seoTitle
          seoDescription
          pricing {
            priceRange {
              start {
                gross {
                  currency
                  amount
                }
              }
              stop {
                gross {
                  currency
                  amount
                }
              }
            }
          }
          media {
            url(size: 2160)
            type
            alt
          }
          collections {
            name
          }
          updatedAt
          variants {
            ...Variant
          }
        }
      }
    }
  }
}
    fragment Variant on ProductVariant {
  id
  name
  attributes {
    attribute {
      slug
      name
      choices(first: 100) {
        edges {
          node {
            name
          }
        }
      }
    }
    values {
      name
    }
  }
  pricing {
    price {
      gross {
        currency
        amount
      }
    }
  }
}`) as unknown as TypedDocumentString<
  GetCollectionProductsBySlugQuery,
  GetCollectionProductsBySlugQueryVariables
>;
// ...
export const GetMenuBySlugDocument = new TypedDocumentString(`
    query GetMenuBySlug($slug: String!) {
  menu(slug: $slug, channel: "default-channel") {
    id
    slug
    name
    items {
      ...MenuItem
      children {
        ...MenuItem
        children {
          ...MenuItem
          children {
            ...MenuItem
          }
        }
      }
    }
  }
}
    fragment MenuItem on MenuItem {
  id
  name
  url
  collection {
    slug
    products(first: 0) {
      totalCount
    }
  }
  category {
    slug
    products(channel: "default-channel", first: 0) {
      totalCount
    }
  }
  page {
    slug
  }
}`) as unknown as TypedDocumentString<GetMenuBySlugQuery, GetMenuBySlugQueryVariables>;
export const GetPageBySlugDocument = new TypedDocumentString(`
    query GetPageBySlug($slug: String!) {
  page(slug: $slug) {
    id
    title
    slug
    content
    seoTitle
    seoDescription
    created
  }
}
    `) as unknown as TypedDocumentString<GetPageBySlugQuery, GetPageBySlugQueryVariables>;
export const GetPagesDocument = new TypedDocumentString(`
    query GetPages {
  pages(first: 10) {
    edges {
      node {
        id
        title
        slug
        content
        seoTitle
        seoDescription
        created
      }
    }
  }
}
    `) as unknown as TypedDocumentString<GetPagesQuery, GetPagesQueryVariables>;
export const GetProductBySlugDocument = new TypedDocumentString(`
    query GetProductBySlug($slug: String!) {
  product(channel: "default-channel", slug: $slug) {
    ...ProductDetails
  }
}
    fragment ProductDetails on Product {
  id
  slug
  name
  isAvailableForPurchase
  description
  seoTitle
  seoDescription
  pricing {
    priceRange {
      start {
        gross {
          currency
          amount
        }
      }
      stop {
        gross {
          currency
          amount
        }
      }
    }
  }
  media {
    url(size: 1080)
    type
    alt
  }
  collections {
    name
  }
  updatedAt
  variants {
    ...Variant
  }
}
fragment Variant on ProductVariant {
  id
  name
  attributes {
    attribute {
      slug
      name
      choices(first: 100) {
        edges {
          node {
            name
          }
        }
      }
    }
    values {
      name
    }
  }
  pricing {
    price {
      gross {
        currency
        amount
      }
    }
  }
}`) as unknown as TypedDocumentString<GetProductBySlugQuery, GetProductBySlugQueryVariables>;